project/requirements.txt
aiogram==3.*
openai==1.*
pytest==8.*
PyYAML==6.*
